Rename cmdline_tools to cmdline
Plus changes to support https://github.com/src-d/lookout-sdk-ml/pull/59
