+++ lookout/style/format/__main__.py
@@ -1,7 +1,7 @@
 """Command line tools entry module."""
 import sys
 
-from lookout.style.format.cmdline_tools import main
+from lookout.style.format.cmdline import main
 
 if __name__ == "__main__":
     sys.exit(main())
