+++ lookout/style/format/cmdline.py
